% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_btns.R
\name{download_btns}
\alias{download_btns}
\alias{download_btns_ui}
\alias{download_btns_server}
\title{Download_btns shiny module.}
\usage{
download_btns_ui(id, settings = list())

download_btns_server(
  id,
  data = reactive({
     NULL
 }),
  name,
  colors = reactive({
     NULL
 }),
  tags = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{internal}

\item{data}{xxxx}

\item{name}{xxxx.}

\item{colors}{xxx}

\item{df.tags}{xxx}
}
\value{
NA
}
\description{
A shiny module that shows download buttons in different formats.
}
