% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error_modal.R
\name{errorModal}
\alias{errorModal}
\alias{errorModal_ui}
\alias{errorModal_server}
\title{Error modal shiny module.}
\usage{
errorModal_ui(id)

errorModal_server(id, text)
}
\arguments{
\item{id}{internal}

\item{text}{xxxx}
}
\value{
NA
}
\description{
A shiny module that shows messages in modal.
}
