% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{find_funs}
\alias{find_funs}
\title{Find the packages of a function}
\usage{
find_funs(f)
}
\arguments{
\item{f}{name of function for which the package(s) are to be identified.}
}
\value{
A dataframe with two columns:
}
\description{
This code is extractd from https://sebastiansauer.github.io/finds_funs/
}
\examples{
find_funs('filter')

}
