% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_format_DT.R
\name{format_DT}
\alias{format_DT}
\alias{format_DT_ui}
\alias{format_DT_server}
\title{format_DT_ui and format_DT_server}
\usage{
format_DT_ui(id)

format_DT_server(id, data)
}
\arguments{
\item{id}{shiny id}

\item{data}{xxx}

\item{withDLBtns}{xxx}

\item{showRownames}{xxx}

\item{dom}{xxx}

\item{hc_style}{xxx}

\item{full_style}{A list of four items:
* cols: a vector of colnames of columns to show,
* vals: a vector of colnames of columns that contain values,
* unique: unique(conds),
* pal: RColorBrewer::brewer.pal(3, "Dark2")[seq_len(2)]}

\item{filename}{xxx}

\item{hideCols}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
