% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert.R
\name{mod_Convert}
\alias{mod_Convert}
\alias{Convert_conf}
\alias{Convert_ui}
\alias{Convert_server}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
Convert_conf()

Convert_ui(id)

Convert_server(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
NA
}
\description{
A shiny Module.
}
\author{
Samuel Wieczorek
}
\keyword{internal}
