% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_insert_md.R
\name{insert_md_ui}
\alias{insert_md_ui}
\alias{insert_md_server}
\title{mod_insert_md_ui and mod_insert_md_server}
\usage{
insert_md_ui(id)

insert_md_server(id, url, link_URL = NULL)
}
\arguments{
\item{id}{shiny id}

\item{url}{internal}
}
\description{
A shiny Module.
}
