% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadapp.R
\name{mod_loading_page}
\alias{mod_loading_page}
\title{mod_loading_page_ui and mod_loading_page_server}
\description{
A shiny Module.
}
\examples{
if(interactive(){
library(shinyjs)
library(shiny)
ui <- mod_loading_page_ui("mod_info")
server <- function(input, output, session) {
  mod_loading_page_server("mod_info")
}

shinyApp(ui, server)
}
}
