% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainapp.R
\name{mod_main_page}
\alias{mod_main_page}
\alias{mainapp_ui}
\alias{mainapp_server}
\title{mod_main_page_ui and mod_loading_page_server}
\usage{
mainapp_ui(id)

mainapp_server(id, funcs = NULL)
}
\arguments{
\item{id}{xxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\examples{
if(interactive()){
library(shinyjs)
library(shiny)
library(shinydashboard)


ui <- mod_main_page_ui("mod_info")

server <- function(input, output, session) {
  mod_main_page_server("mod_info")
}

shinyApp(ui, server)

}

}
\keyword{internal}
