% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataManager.R
\name{dataManager_ui}
\alias{dataManager_ui}
\alias{dataManager_server}
\title{mod_open_demo_dataset_ui and mod_open_demo_dataset_server}
\usage{
dataManager_ui(id)

dataManager_server(id, funcs)
}
\arguments{
\item{id}{shiny id}

\item{pipeline.def}{xxx}
}
\value{
An object of class [`xxxx`]
}
\description{
A shiny Module.
}
\keyword{internal}
