% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_view_dataset.R
\name{view_dataset_dataset}
\alias{view_dataset_dataset}
\alias{view_dataset_ui}
\title{mod_open_demo_dataset_ui and mod_open_demo_dataset_server}
\usage{
view_dataset_ui(id)
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module.
}
\keyword{internal}
