% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_mod_insert_md.R
\name{generic_mod_insert_md}
\alias{generic_mod_insert_md}
\alias{default_mod_insert_md_ui}
\alias{default_mod_insert_md_server}
\title{mod_insert_md_ui and mod_insert_md_server}
\usage{
default_mod_insert_md_ui(id)

default_mod_insert_md_server(id, url)
}
\arguments{
\item{id}{shiny id}

\item{url}{xxx}
}
\description{
A shiny Module.
}
\keyword{internal}
