% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_demoDataset.R
\name{generic_mod_open_demo_dataset}
\alias{generic_mod_open_demo_dataset}
\alias{open_demoDataset_ui}
\alias{open_demoDataset_server}
\title{mod_open_demo_dataset_ui and mod_open_demo_dataset_server}
\usage{
open_demoDataset_ui(id)

open_demoDataset_server(id)
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module.
}
\keyword{internal}
