% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_mod_convert.R
\name{mod_Convert}
\alias{mod_Convert}
\alias{default_Convert_conf}
\alias{default_Convert_ui}
\alias{default_Convert_server}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
default_Convert_conf()

default_Convert_ui(id)

default_Convert_server(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
NA
}
\description{
A shiny Module.
}
\author{
Samuel Wieczorek
}
\keyword{internal}
