% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_mod_convert.R
\name{mod_Convert}
\alias{mod_Convert}
\alias{generic_Convert_conf}
\alias{generic_Convert_ui}
\alias{generic_Convert_server}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
generic_Convert_conf()

generic_Convert_ui(id)

generic_Convert_server(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
NA
}
\description{
A shiny Module.
}
\author{
Samuel Wieczorek
}
\keyword{internal}
