% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert.R
\name{mod_Convert}
\alias{mod_Convert}
\alias{convert_conf}
\alias{convert_ui}
\alias{convert_server}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
convert_conf()

convert_ui(id)

convert_server(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
NA
}
\description{
A shiny Module.
}
\author{
Samuel Wieczorek
}
\keyword{internal}
